implemented paddle controller
reworked input event system to better support paddle input

updated recording/playback fileformat to support new input event system

updated setup package to support new input event system

updated debugger commands to support new input event system

added mouse capture request to GUI interface - in SDL GUI
implementations, paddle input only works once mouse has been captured

changed gui/requests to match coding style used in input/events

+++ gui/events.go
@@ -48,8 +48,9 @@ type EventKeyboard struct {
 
 // EventMouseMotion is the data that accompanies MouseEventMove events
 type EventMouseMotion struct {
-	X int
-	Y int
+	// as a fraction of the window's dimensions
+	X float32
+	Y float32
 }
 
 // MouseButton identifies the mouse button
